android-versie/requirements-android.txt
# Android/Termux Optimized Requirements for Agent Zero
# Versie: 1.1 - November 28, 2025

# Core AI/LLM Libraries
anthropic>=0.69.0
openai>=2.8.0
langchain-anthropic>=0.1.19
langchain-openai>=0.1.15
langchain-community>=0.2.7

# Active LLM Providers
langchain-google-genai>=1.0.7
langchain-groq>=0.1.6
# langchain-mistralai>=0.1.8
# langchain-ollama>=0.1.3

# Embeddings (lightweight versie)
# sentence-transformers>=3.0.1  # HEAVY: Requires PyTorch. Uncomment only if you need local embeddings.
# langchain-huggingface>=0.0.3

# Web & Network
beautifulsoup4>=4.12.3
duckduckgo-search>=6.1.12
lxml-html-clean>=0.2.0
# newspaper3k==0.2.8  # Kan problemen geven, skip indien niet nodig

# SSH & Remote (voor code execution)
paramiko>=3.4.0

# Web Framework (voor UI)
flask>=3.0.3
flask-basicauth>=0.2.0

# Utilities
python-dotenv>=1.0.1
markdown>=3.7
pypdf>=4.3.1
ansio>=0.0.1
inputimeout>=1.0.4
webcolors>=24.6.0

# SKIP op Android (te zwaar of niet compatibel):
# faiss-cpu  # Te groot voor compile op ARM - gebruik alternatief
# docker==7.1.0  # Niet nodig zonder Docker execution
# unstructured==0.15.13  # Te veel dependencies
# unstructured-client==0.25.9  # Niet nodig

# Lightweight Vector Store Alternatief (optioneel)
# chromadb  # Lichter dan faiss, werkt goed op ARM
